requirements.txt
fastapi
uvicorn
pdfplumber
python-multipart
google-generativeai
slowapi